Safari-611.1.10/LayoutTests/http/tests/svg/svg-use-external.html

<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">

    <title>Verify that the SVG use resource only loads once</title>
    <script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpResourceLoadCallbacks();
        testRunner.waitUntilDone();
        onload = function() {
            testRunner.notifyDone();
        }
    }
    </script>
</head>
<body>
    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-rocket"></use>
    </svg>

    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-fire"></use>
    </svg>

    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-lab"></use>
    </svg>

    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-magnet"></use>
    </svg>

    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-shield"></use>
    </svg>

    <svg>
        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="resources/symbol-defs.svg#icon-power"></use>
    </svg>
</body>
</html>
