Safari-611.1.10/LayoutTests/http/tests/svg/svg-use-external-expected.txt

http://127.0.0.1:8000/svg/resources/symbol-defs.svg#icon-rocket - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/svg/resources/symbol-defs.svg#icon-rocket, main document URL http://127.0.0.1:8000/svg/svg-use-external.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/svg/svg-use-external.html - didFinishLoading
http://127.0.0.1:8000/svg/resources/symbol-defs.svg#icon-rocket - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/svg/resources/symbol-defs.svg#icon-rocket, http status code 200>
http://127.0.0.1:8000/svg/resources/symbol-defs.svg#icon-rocket - didFinishLoading
